gems/requirements.txt

Django==2.2.14
djangorestframework==3.11.0
